+++ packages/atomic/src/components/common/facets/facet-container/facet-container.ts
@@ -1,0 +1,12 @@
+import {FunctionalComponentWithChildrenNoProps} from '@/src/utils/functional-component-utils';
+import {html} from 'lit';
+
+export const renderFacetContainer: FunctionalComponentWithChildrenNoProps =
+  () => (children) => html`
+    <div
+      class="bg-background border-neutral rounded-lg border p-4"
+      part="facet"
+    >
+      ${children}
+    </div>
+  `;